STEP 10. Break up search results events.
