+++ altitude/static/js/fragments/index.js
@@ -2,6 +2,7 @@ import { findFragmentRoots } from "./helpers.js"
 import { hydrateModalFragment } from "./modal.js"
 import { hydrateImageDetailFragment } from "./image-detail.js"
 import { hydratePersonNameEditorFragment } from "./person-name-editor.js"
+import { hydrateSearchResultsFragment } from "./search-results.js"
 
 export function hydrateAppFragments({ root, app }) {
     findFragmentRoots(root, "person-name-editor").forEach((fragmentEl) => {
@@ -26,7 +27,7 @@ export function hydrateAppFragments({ root, app }) {
     })
 
     findFragmentRoots(root, "search-results").forEach((fragmentEl) => {
-        app.hydrateSearchResultsFragment(fragmentEl)
+        hydrateSearchResultsFragment({ fragmentEl, app })
     })
 }
