+++ altitude/static/js/frontend-app.js
@@ -12,12 +12,10 @@ import {
 } from "./common/dragon-drop.js"
 import { showAssetDetailModal } from "./common/modal.js"
 import { hydrateAppFragments } from "./fragments/index.js"
+import { handleViewSettingChanged } from "./fragments/search-results.js"
 import { registerAppEventListeners } from "./listeners/index.js"
 import { setImgSrcAndWait } from "./search-results/detail-navigator.js"
 import "./search-results/dragon-drop.js"
-
-const placeholderImageData =
-    "data:image/gif;base64,R0lGODlhAQABAIAAAP///wAAACH5BAEAAAAALAAAAAABAAEAAAICRAEAOw=="
 
 export class FrontendApp {
     constructor({ Alpine, context }) {
@@ -579,187 +577,8 @@ export class FrontendApp {
         }
     }
 
-
-    hydrateSearchResultsFragment(fragmentEl) {
-        const resultsTotal = Number(fragmentEl.dataset.resultsTotal || 0)
-        const contentElement = document.getElementById("content")
-        const assetsElement = fragmentEl.querySelector("#assets")
-
-        if (contentElement) {
-            contentElement.scrollTo({ top: 0, behavior: "auto" })
-        }
-
-        this.Alpine.store(Const.state.selectedAssets).reset()
-        this.Alpine.store(Const.state.resultsTotal).set(resultsTotal)
-        this.Alpine.store(Const.state.shadowResults).reset()
-
-        if (!assetsElement) {
-            return
-        }
-
-        this.bindSearchResultsInfiniteScroll(assetsElement)
-        this.bindSearchResultsLazyLoad(assetsElement)
-        this.applyGridMetadataVisibilityToAllCells(assetsElement)
-        this.syncShadowResultsFromSearchUrl()
-    }
-
-    bindSearchResultsInfiniteScroll(assetsElement) {
-        if (assetsElement.dataset.appInfiniteScrollBound === "true") {
-            return
-        }
-
-        assetsElement.dataset.appInfiniteScrollBound = "true"
-
-        assetsElement.addEventListener("htmx:beforeRequest", (event) => {
-            if (!event.target.classList.contains("last-cell")) {
-                return
-            }
-
-            if (event.detail.target.getAttribute("data-hx-revealed")) {
-                event.preventDefault()
-            } else {
-                console.debug(
-                    "Loading more: %s",
-                    event.detail.pathInfo.requestPath,
-                )
-            }
-        })
-
-        assetsElement.addEventListener("htmx:afterRequest", (event) => {
-            if (!event.target.classList.contains("last-cell")) {
-                return
-            }
-
-            event.detail.target.setAttribute("data-hx-revealed", "true")
-
-            if (event.detail.successful) {
-                this.appendShadowResultsForRequestPath(
-                    event.detail.pathInfo.requestPath,
-                )
-            }
-        })
-    }
-
-    bindSearchResultsLazyLoad(assetsElement) {
-        if (assetsElement.dataset.appLazyLoadBound === "true") {
-            return
-        }
-
-        assetsElement.dataset.appLazyLoadBound = "true"
-
-        const observer = this.getLazyImageObserver()
-
-        assetsElement.addEventListener("htmx:load", (event) => {
-            const imgEl = event.target.querySelector("img")
-            if (imgEl) {
-                observer.observe(imgEl)
-            }
-
-            this.showOrHideAssetGridMetadata(event.target)
-        })
-
-        assetsElement.querySelectorAll(".cell").forEach((cellEl) => {
-            const imgEl = cellEl.querySelector("img")
-            if (imgEl) {
-                observer.observe(imgEl)
-            }
-
-            this.showOrHideAssetGridMetadata(cellEl)
-        })
-    }
-
-    getLazyImageObserver() {
-        if (this.lazyImageObserver) {
-            return this.lazyImageObserver
-        }
-
-        const observerOptions = {
-            root: null,
-            rootMargin: "0px 100% 0px 100%",
-            threshold: [0, 1],
-        }
-
-        this.lazyImageObserver = new IntersectionObserver((entries) => {
-            entries.forEach((entry) => {
-                const imgEl = entry.target
-
-                if (entry.isIntersecting) {
-                    if (
-                        imgEl.getAttribute("src") ===
-                        imgEl.getAttribute(Const.attributes.dataSrc)
-                    ) {
-                        return
-                    }
-
-                    if (imgEl.hasAttribute(Const.attributes.dataSrc)) {
-                        imgEl.src = imgEl.getAttribute(Const.attributes.dataSrc)
-                    }
-
-                    return
-                }
-
-                if (imgEl.getAttribute("src") === placeholderImageData) {
-                    return
-                }
-
-                if (imgEl.hasAttribute(Const.attributes.dataSrc)) {
-                    imgEl.src = placeholderImageData
-                }
-            })
-        }, observerOptions)
-
-        return this.lazyImageObserver
-    }
-
-    applyGridMetadataVisibilityToAllCells(assetsElement) {
-        assetsElement.querySelectorAll(".cell").forEach((cellEl) => {
-            this.showOrHideAssetGridMetadata(cellEl)
-        })
-    }
-
-    showOrHideAssetGridMetadata(cellEl) {
-        const showFields = this.context.getGridMetadataFields()
-        const metadata = cellEl.querySelector(".metadata")
-
-        if (!metadata) {
-            return
-        }
-
-        const toShowFieldsSelectorStr = Array.from(showFields)
-            .map((fieldName) => `.metadata > div.${fieldName}`)
-            .join(", ")
-
-        metadata.style.display = toShowFieldsSelectorStr.length ? "grid" : "none"
-
-        if (!toShowFieldsSelectorStr.length) {
-            return
-        }
-
-        cellEl.querySelectorAll(".metadata > div").forEach((divEl) => {
-            divEl.style.display = "none"
-        })
-
-        cellEl.querySelectorAll(toShowFieldsSelectorStr).forEach((divEl) => {
-            divEl.style.display = "block"
-        })
-    }
-
     handleViewSettingChanged(event) {
-        const fieldName = event.detail["fieldName"]
-        const checked = event.detail["checked"]
-
-        document
-            .querySelectorAll(`.metadata > div.${fieldName}`)
-            .forEach((divEl) => {
-                divEl.style.display = checked ? "block" : "none"
-            })
-
-        const showFields = this.context.getGridMetadataFields()
-        const metadataDisplay = showFields.size === 0 ? "none" : "grid"
-
-        document.querySelectorAll("#assets .metadata").forEach((divEl) => {
-            divEl.style.display = metadataDisplay
-        })
+        handleViewSettingChanged({ event, context: this.context })
     }
 
     syncShadowResultsFromSearchUrl() {
